CB-5299: Cache pbxproj to avoid re-parsing it for each plugin.
Parsing pbxproj is very slow (~250ms on my machine). It is done for each plugin
that needs to add frameworks to the ios project. This change is a quick and
dirty fix to allow caching the parsed pbxproj object so that it can be reused for
multiple plugins during prepare.

Timings on a test project for cordova prepare ios:
before: 1.623s
after:  0.981s